knowledge/requirements.txt
numpy==1.24.3
faiss-cpu==1.7.4
pydantic==2.0.0
python-dotenv==1.0.0
